Infra update to add preemtible nodes

+++ cloudsql/indexdb.sql
@@ -7,7 +7,7 @@ CREATE TABLE IF NOT EXISTS event_index (
     id SERIAL PRIMARY KEY,
     filename TEXT,
     timestamp BIGINT,   -- ms since epoch
-    start_index INTEGER,
-    end_index INTEGER,
+    start_index BIGINT,
+    end_index BIGINT,
     event_id SERIAL REFERENCES events(id)
 );
