+++ tweets-api/src/main/java/edu/colorado/cs/epic/tweetsapi/core/EventRow.java
@@ -3,8 +3,8 @@ package edu.colorado.cs.epic.tweetsapi.core;
 public class EventRow {
     String filename;
     long timestamp;
-    int startIndex;
-    int endIndex;
+    long startIndex;
+    long endIndex;
 
     public EventRow() {
         filename = "";
@@ -13,7 +13,7 @@ public class EventRow {
         endIndex = 0;
     }
 
-    public EventRow(String filename, long timestamp, int startIndex, int endIndex) {
+    public EventRow(String filename, long timestamp, long startIndex, long endIndex) {
         this.filename = filename;
         this.timestamp = timestamp;
         this.startIndex = startIndex;
@@ -35,19 +35,19 @@ public class EventRow {
     public long getTimestamp() {
         return timestamp;
     }
-    public void setStartindex(int startIndex) {
+    public void setStartindex(long startIndex) {
         this.startIndex = startIndex;
     }
 
-    public int getStartIndex() {
+    public long getStartIndex() {
         return startIndex;
     }
 
-    public void setEndIndex(int endIndex) {
+    public void setEndIndex(long endIndex) {
         this.endIndex = endIndex;
     }
 
-    public int getEndIndex() {
+    public long getEndIndex() {
         return endIndex;
     }
 }
